Move dispatch function to ewm
