+++ pandas/core/window/rolling.py
@@ -506,22 +506,6 @@ class BaseWindow(ShallowMixin, SelectionMixin):
     agg = aggregate
 
 
-def dispatch(name: str, *args, **kwargs):
-    """
-    Dispatch to groupby apply.
-    """
-
-    def outer(self, *args, **kwargs):
-        def f(x):
-            x = self._shallow_copy(x, groupby=self._groupby)
-            return getattr(x, name)(*args, **kwargs)
-
-        return self._groupby.apply(f)
-
-    outer.__name__ = name
-    return outer
-
-
 class BaseWindowGroupby(GotItemMixin, BaseWindow):
     """
     Provide the groupby windowing facilities.
